Improvement to #43
Not a bad idea to display artist name, and album name, but I think it is
useless to return the basename each time mpd could not find audio tags.
Returning "unknown artist" or "unknown album" is better for me. I keep
the basename return in title because it is important.