Added support for server side plugins
Running sh script and executing binaries is not the best way to make a
plugin system, but in my case, my needs were to be able to launch
command line file downloader to add music to mpd ( adding a stream
doesn't work ). This plugin system is capable of running multiple thread
in parallel in order to let the ympd server run without interruption.